feat(import/anytype): import code blocks with language preserved
Anytype Code-style blocks become Trilium/CKEditor code blocks
(<pre><code class="language-...">), mapping the PrismJS language id in
fields.lang to a Trilium MIME via getMimeTypeFromMarkdownName (with a
clike->c alias for PrismJS's C-family base); unknown languages fall back
to auto-detect. Quotes are kept literal, matching the markdown importer.

Also consolidates the importer back into a single module: the model
types and the marks/code rendering (previously split into model.ts,
marks.ts, code.ts, html.ts) now live in importer.ts, and rendering reuses
the shared escapeHtml from utils instead of a local copy.
