+++ packages/trilium-core/src/services/import/anytype/importer.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
 
-import { isPage, parseObject } from "./importer.js";
-import type { AnytypeBlock, AnytypeMark, AnytypeSnapshot } from "./model.js";
+import { isPage, parseObject, renderCodeBlock, renderInlineText } from "./importer.js";
+import type { AnytypeBlock, AnytypeMark, AnytypeSnapshot } from "./importer.js";
 
 /** Wraps blocks + details into the export's snapshot shape. */
 function snapshot(blocks: AnytypeBlock[], details: { id?: string; name?: string; layout?: number; resolvedLayout?: number }, sbType = "Page"): AnytypeSnapshot {
@@ -26,6 +26,11 @@ function page(name: string, contentBlocks: AnytypeBlock[], layout = 0): AnytypeS
     );
 }
 
+/** An inline mark over a `[from, to)` range. */
+function mark(from: number, to: number, type: string, param?: string): AnytypeMark {
+    return { range: { from, to }, type, param };
+}
+
 describe("isPage", () => {
     it("accepts a basic-layout Page and rejects sets and system objects", () => {
         expect(isPage(page("A page", []))).toBe(true);
@@ -48,7 +53,7 @@ describe("isPage", () => {
 
 describe("parseObject", () => {
     it("takes the title from details.name and emits non-heading text blocks as paragraphs", () => {
-        // Numbered/Marked/Quote/Code etc. are all flattened to <p> for now (only headings get their own tag).
+        // Numbered/Marked/Quote etc. are all flattened to <p> for now (only headings get their own tag).
         const result = parseObject(page("My Page", [textBlock("b1", "First"), textBlock("b2", "Second", "Numbered")]));
         expect(result.id).toBe("obj");
         expect(result.title).toBe("My Page");
@@ -69,8 +74,14 @@ describe("parseObject", () => {
     });
 
     it("applies inline marks inside the block's tag", () => {
-        const result = parseObject(page("Marks", [textBlock("b1", "Bold text", "Paragraph", [], [{ range: { from: 0, to: 4 }, type: "Bold" }])]));
+        const result = parseObject(page("Marks", [textBlock("b1", "Bold text", "Paragraph", [], [mark(0, 4, "Bold")])]));
         expect(result.content).toBe("<p><strong>Bold</strong> text</p>");
+    });
+
+    it("renders a Code-style block as a code block, preserving the language from fields.lang", () => {
+        const codeBlock: AnytypeBlock = { id: "b1", text: { text: "int x;", style: "Code" }, fields: { lang: "clike" }, childrenIds: [] };
+        const result = parseObject(page("Code", [codeBlock]));
+        expect(result.content).toBe('<pre><code class="language-text-x-csrc">int x;</code></pre>');
     });
 
     it("walks nested blocks in document order (parent text before its children)", () => {
@@ -89,13 +100,111 @@ describe("parseObject", () => {
         expect(result.content).toBe("<p>Real content</p>");
     });
 
-    it("escapes HTML special characters in the text", () => {
-        const result = parseObject(page("Escaping", [textBlock("b1", "a < b & c > d")]));
-        expect(result.content).toBe("<p>a &lt; b &amp; c &gt; d</p>");
-    });
-
     it("falls back to 'Untitled' when the page has no name", () => {
         expect(parseObject(page("", [textBlock("b1", "body")])).title).toBe("Untitled");
         expect(parseObject(page("   ", [])).title).toBe("Untitled");
+    });
+});
+
+describe("renderInlineText", () => {
+    it("returns escaped plain text when there are no marks", () => {
+        expect(renderInlineText("a < b & c > d", [])).toBe("a &lt; b &amp; c &gt; d");
+    });
+
+    it("wraps a single mark's range, leaving the rest untouched", () => {
+        expect(renderInlineText("Bold text", [mark(0, 4, "Bold")])).toBe("<strong>Bold</strong> text");
+    });
+
+    it("maps the five supported marks to their tags ([from, to) range)", () => {
+        expect(renderInlineText("x", [mark(0, 1, "Bold")])).toBe("<strong>x</strong>");
+        expect(renderInlineText("x", [mark(0, 1, "Italic")])).toBe("<em>x</em>");
+        expect(renderInlineText("x", [mark(0, 1, "Strikethrough")])).toBe("<s>x</s>");
+        expect(renderInlineText("x", [mark(0, 1, "Underscored")])).toBe("<u>x</u>");
+        expect(renderInlineText("x", [mark(0, 1, "Keyboard")])).toBe("<code>x</code>");
+    });
+
+    it("renders the real 'Formatting test' line, nesting coincident bold+italic+underline", () => {
+        // Verbatim text and marks from the exported page (marks intentionally unsorted, as in the export).
+        const text = "Bold Italic Strikethrough Underline Bold Italic Underline";
+        const marks = [
+            mark(12, 25, "Strikethrough"),
+            mark(5, 11, "Italic"),
+            mark(36, 57, "Italic"),
+            mark(0, 4, "Bold"),
+            mark(36, 57, "Bold"),
+            mark(26, 35, "Underscored"),
+            mark(36, 57, "Underscored")
+        ];
+        expect(renderInlineText(text, marks)).toBe(
+            "<strong>Bold</strong> <em>Italic</em> <s>Strikethrough</s> <u>Underline</u> <strong><em><u>Bold Italic Underline</u></em></strong>"
+        );
+    });
+
+    it("splits partially overlapping marks into properly nested segments", () => {
+        // Bold [0,4) and Italic [2,6) overlap only on [2,4).
+        expect(renderInlineText("abcdef", [mark(0, 4, "Bold"), mark(2, 6, "Italic")])).toBe(
+            "<strong>ab</strong><strong><em>cd</em></strong><em>ef</em>"
+        );
+    });
+
+    it("maps TextColor to a colour span and a BackgroundColor highlight to a colour + background span", () => {
+        expect(renderInlineText("Red", [mark(0, 3, "TextColor", "red")])).toBe('<span style="color:#e2400c">Red</span>');
+        // A highlight with no explicit text colour gets Anytype's default dark text (#252525) so it stays
+        // readable on dark themes — otherwise the theme-default white text is invisible on the pale highlight.
+        expect(renderInlineText("Red", [mark(0, 3, "BackgroundColor", "red")])).toBe(
+            '<span style="color:#252525;background-color:#fcd1c3">Red</span>'
+        );
+    });
+
+    it("combines co-occurring text + background colour into one span, nested inside structural marks", () => {
+        // An explicit text colour wins over the highlight default, and both fold into a single span.
+        expect(renderInlineText("Red", [mark(0, 3, "TextColor", "red"), mark(0, 3, "BackgroundColor", "red")])).toBe(
+            '<span style="color:#e2400c;background-color:#fcd1c3">Red</span>'
+        );
+        // A structural mark stays outermost; the colour span nests inside it.
+        expect(renderInlineText("Red", [mark(0, 3, "Bold"), mark(0, 3, "TextColor", "red")])).toBe(
+            '<strong><span style="color:#e2400c">Red</span></strong>'
+        );
+    });
+
+    it("colours each word independently across a palette line (as the page exports it)", () => {
+        expect(renderInlineText("Grey Red", [mark(0, 4, "TextColor", "grey"), mark(5, 8, "TextColor", "red")])).toBe(
+            '<span style="color:#8c9ea5">Grey</span> <span style="color:#e2400c">Red</span>'
+        );
+    });
+
+    it("ignores an unknown colour name and genuinely unsupported mark types", () => {
+        expect(renderInlineText("x", [mark(0, 1, "TextColor", "chartreuse")])).toBe("x");
+        expect(renderInlineText("@bob", [mark(0, 4, "Mention", "someObjectId")])).toBe("@bob");
+    });
+
+    it("escapes HTML inside a marked range", () => {
+        expect(renderInlineText("a<b", [mark(0, 3, "Bold")])).toBe("<strong>a&lt;b</strong>");
+    });
+
+    it("clamps out-of-range offsets and drops empty / reversed ranges", () => {
+        expect(renderInlineText("hi", [mark(0, 100, "Bold")])).toBe("<strong>hi</strong>");
+        expect(renderInlineText("hi", [mark(1, 1, "Bold")])).toBe("hi");
+        expect(renderInlineText("hi", [mark(2, 0, "Bold")])).toBe("hi");
+    });
+});
+
+describe("renderCodeBlock", () => {
+    it("wraps code in <pre><code> with the resolved language class", () => {
+        // Anytype tags C-family code as PrismJS "clike"; Trilium has no such code, so it's aliased to C.
+        expect(renderCodeBlock("int x;", "clike")).toBe('<pre><code class="language-text-x-csrc">int x;</code></pre>');
+    });
+
+    it("resolves a language code that matches a Trilium markdown name directly", () => {
+        expect(renderCodeBlock("print(1)", "python")).toBe('<pre><code class="language-text-x-python">print(1)</code></pre>');
+    });
+
+    it("falls back to auto-detect for an unknown or missing language", () => {
+        expect(renderCodeBlock("plain", "nonsense")).toBe('<pre><code class="language-text-x-trilium-auto">plain</code></pre>');
+        expect(renderCodeBlock("plain", undefined)).toBe('<pre><code class="language-text-x-trilium-auto">plain</code></pre>');
+    });
+
+    it("escapes HTML and keeps quotes literal, preserving newlines and tabs", () => {
+        expect(renderCodeBlock('a<b & c>\n\t"q"', "clike")).toBe('<pre><code class="language-text-x-csrc">a&lt;b &amp; c&gt;\n\t"q"</code></pre>');
     });
 });
